@@ -7,11 +7,11 @@ description: Unofficial Amazon Cognito Identity Provider Dart SDK,
   with AWS Cloud Services.
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  crypto: ^3.0.0
-  http: ^0.13.1
+  crypto: ^3.0.1
+  http: ^0.13.3
 
 dev_dependencies:
   flutter_test:
